v1/requirements.txt
# Core
python-dotenv>=1.0.0
requests>=2.31.0

# LiveKit Agents Framework + Sarvam plugin
livekit-agents[sarvam,openai,silero]>=1.0.0
livekit-plugins-sarvam>=1.0.0

# OpenAI (for LLM conversation + extraction)
openai>=1.30.0

# Sarvam AI SDK (for standalone API testing)
sarvamai>=0.1.10

# Excel export
openpyxl>=3.1.0

# Testing
pytest>=7.4
pytest-asyncio>=0.21
